+++ .env
@@ -1,3 +1,4 @@
 #PUBLIC_BASE_URL=https://server-alpha.tech
-PUBLIC_BASE_URL=http://localhost:3000
+# PUBLIC_BASE_URL=http://localhost:3000
+PUBLIC_BASE_URL=http://0.0.0.0:3000
 #PUBLIC_BASE_URL=https://server-beta.de
